requirements.txt
PyQt6==6.6.1
PyQt6-Qt6==6.6.1
PyQt6-sip==13.6.0
python-dateutil==2.8.2
Pillow==10.1.0
pyinstaller==6.1.0
